+++ index.html
@@ -131,7 +131,6 @@
         <span class="popup__title popup__title_img"></span>
       </div>
     </div>
-    <script type="module" src="./scripcard.js"></script>
-    <script type="module" src="./script.js"></script>
+    <script type="module" src="./script/index.js"></script>
   </body>
 </html>
